BuildTemplateLink id wrappers

+++ v2-model/src/commonMain/kotlin/com/bselzer/gw2/v2/model/character/specialization/CharacterSpecialization.kt
@@ -21,4 +21,4 @@ data class CharacterSpecialization(
      */
     @SerialName("traits")
     val traits: List<TraitId?> = emptyList()
-) : Identifiable<Int>
+) : Identifiable<Byte>
